requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.2.1
graphene==3.2.2
graphene-django==3.0.2
graphene-federation==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
django-oauth2==4.0
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-auditlog==2.2.2 
python-dateutil==2.8.2
requests==2.31.0
pyasn1==0.5.0 
pyasn1-modules==0.3.0
graphene-file-upload==1.3.0
python-magic==0.4.27
jinja2
lxml==4.9.2 
python-docx==0.8.11
pillow==10.0.1